Placed classes in the appropriate packages
