+++ WebShop/src/main/java/com/tms/webshop/Filters/AuthFilter.java
@@ -1,4 +1,4 @@
-package com.tms.webshop;
+package com.tms.webshop.Filters;
 
 import jakarta.servlet.*;
 import jakarta.servlet.annotation.*;
