+++ WebShop/src/main/java/com/tms/webshop/servlets/RemoveProductServlet.java
@@ -1,4 +1,4 @@
-package com.tms.webshop;
+package com.tms.webshop.servlets;
 
 import jakarta.servlet.*;
 import jakarta.servlet.http.*;
@@ -7,13 +7,17 @@ import jakarta.servlet.annotation.*;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.stream.Stream;
 
-@WebServlet(name = "AddProductServlet", value = "/add-product")
-public class AddProductServlet extends HttpServlet {
+@WebServlet(name = "RemoveProductServlet", value = "/remove-product")
+public class RemoveProductServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         HashMap<Integer, Integer> basketProductsMap = (HashMap<Integer, Integer>) request.getSession().getAttribute("basketProductsMap");
-        basketProductsMap.merge(Integer.parseInt(request.getParameter("productId")), 1, Integer::sum);
+        Integer productId = Integer.parseInt(request.getParameter("productId"));
+        if (basketProductsMap.get(productId) > 1) {
+            basketProductsMap.compute(productId, (key, value) -> value - 1);
+        } else {
+            basketProductsMap.remove(productId);
+        }
     }
 }
