+++ WebShop/src/main/java/com/tms/webshop/utilsDB/DBConnectionContainer.java
@@ -1,4 +1,4 @@
-package com.tms.utilsDB;
+package com.tms.webshop.utilsDB;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
